Fix breakage on MacOs + Chrome 94
On Chrome 94 on MacOs, canvas drawing is broken. I traced this back to
the 'alpha: false' setting on the offscreen game canvas. No idea why
this causes issues; can only assume there is a bug either in Chrome or
MacOS (all other combinations of browser + OS seem to be working
fine).