Add zombies
Zombies are 2D ragdolls that interact with the canvas via physics
simulation. They were a feature of some of the original sand falling
games, and thus were something I've long desired to add.

I've attempted to replicate most of the original zombie behavior, while
also making some alterations and additions. The size and movement of the
zombies differs somewhat from the original, and I have also added some
addition features like the interaction between zombies and ice.

Similar to the particle implementation (scripts/particles.js), the
soft body simulation is done in a separate context from the main game
canvas, and then copied over for each frame.

Zombies are just one type of physics soft body that can be implemented.
In the future, other types of soft body simulations can be added to
the game. matter.js provides a great deal of flexibility in terms of
the simulation possibilities.